out with .ini, in with .json
Removed all the files pertaining to dandere2x that used .ini files and is now entirely based on json
